Remote refreshing and new folder fixed
