+++ main/utils/remoteFunctions.js
@@ -1,5 +1,10 @@
 import { Client } from "ssh2"
+import { app, ipcMain } from "electron"
+import { mainWindow } from "../background"
+import { generateSSHKeyPair } from '../sshKeygen'
 const net = require("net")
+var path = require("path")
+const fs = require("fs")
 
 // Global tunnel state for remote connection management
 let activeTunnel = null
@@ -31,20 +36,70 @@ export function getRemoteWorkspacePath() {
   return remoteWorkspacePath
 }
 
+// Tunnel information and state management
+let tunnelInfo = {
+  host: null,
+  tunnelActive: false,
+  localAddress: "localhost",
+  localBackendPort: null,
+  remoteBackendPort: null,
+  localDBPort: null,
+  remoteDBPort: null,
+  remotePort: null,
+  username: null,
+};
+
+export function setTunnelState(info) {
+  console.log("Setting tunnel state with info:", info);
+  tunnelInfo = { ...tunnelInfo, ...info, tunnelActive: info.tunnelActive }
+}
+
+export function clearTunnelState() {
+  tunnelInfo = {
+    host: null,
+    tunnelActive: false,
+    localAddress: "localhost",
+    localBackendPort: null,
+    remoteBackendPort: null,
+    localDBPort: null,
+    remoteDBPort: null,
+    remotePort: null,
+    username: null,
+  };
+}
+
+export function getTunnelState() {
+  console.log("Getting tunnel state:", tunnelInfo);
+  return tunnelInfo
+}
+
+ipcMain.handle('getTunnelState', () => {
+  return getTunnelState()
+})
+
+ipcMain.handle('setTunnelState', (_event, info) => {
+  setTunnelState(info)
+  mainWindow.webContents.send('tunnelStateUpdated', info);
+})
+
+ipcMain.handle('clearTunnelState', () => {
+  clearTunnelState()
+})
+
 
 /**
  * Starts an SSH tunnel and creates the backend port forwarding server only.
  * MongoDB tunnel can be created later by calling startMongoTunnel.
  * @param {Object} params - SSH and port config.
- * @param {string} params.host
- * @param {string} params.username
- * @param {string} [params.privateKey]
- * @param {string} [params.password]
- * @param {number|string} params.remotePort
- * @param {number|string} params.localBackendPort
- * @param {number|string} params.remoteBackendPort
- * @param {number|string} params.localDBPort
- * @param {number|string} params.remoteDBPort
+ * @param {string} params.host - Address of the remote host.
+ * @param {string} params.username - Username for SSH connection.
+ * @param {string} [params.privateKey] - Private key for SSH authentication.
+ * @param {string} [params.password] - Password for SSH authentication.
+ * @param {number|string} params.remotePort - Port of the SSH connection
+ * @param {number|string} params.localBackendPort - Local port forwarded to the remote backend.
+ * @param {number|string} params.remoteBackendPort - Port on the remote host for the backend server.
+ * @param {number|string} params.localDBPort - Local port for the MongoDB server.
+ * @param {number|string} params.remoteDBPort - Port on the remote host for the MongoDB server.
  * @returns {Promise<{success: boolean}>}
  */
 export async function startSSHTunnel({ host, username, privateKey, password, remotePort, localBackendPort, remoteBackendPort, localDBPort, remoteDBPort }) {
@@ -148,7 +203,7 @@ async function checkRemotePortOpen(conn, port) {
 }
 
 /**
- * Starts the MongoDB port forwarding tunnel using an existing SSH connection.
+ * @description Starts the MongoDB port forwarding tunnel using an existing SSH connection.
  * Checks if the remote port is open before creating the tunnel, with retries.
  * @returns {Promise<{success: boolean}>}
  */
@@ -217,7 +272,7 @@ export async function startMongoTunnel() {
 }
 
 /**
- * Confirms that the mongoDB tunnel is active and the server is listening.
+ * @description Confirms that the mongoDB tunnel is active and the server is listening.
  * @returns {Promise<{success: boolean, error?: string}>}
  */
 export async function confirmMongoTunnel() {
@@ -243,7 +298,7 @@ export async function confirmMongoTunnel() {
 }
 
 /**
- * Stops the SSH tunnel and closes all forwarded servers.
+ * @description Stops the SSH tunnel and closes all forwarded servers.
  * @returns {Promise<{success: boolean, error?: string}>}
  */
 export async function stopSSHTunnel() {
@@ -280,37 +335,11 @@ export async function stopSSHTunnel() {
   return { success: false, error: error || "No active tunnel" }
 }
 
-export function checkRemoteFolderExists(folderPath) {
-  // Ensure tunnel is active and SSH client is available
-  const tunnelObject = getActiveTunnel()
-  if (!tunnelObject) {
-    const errMsg = "No active SSH tunnel for remote folder creation."
-    console.error(errMsg)
-    return Promise.resolve("tunnel inactive")
-  }
-
-  return new Promise((resolve, reject) => {
-    tunnelObject.sftp((err, sftp) => {
-      if (err) {
-        console.error("SFTP error:", err)
-        resolve("sftp error")
-        return
-      }
-
-      // Check if folder exists
-      sftp.stat(folderPath, (statErr, stats) => {
-        if (!statErr && stats && stats.isDirectory()) {
-          // Folder exists
-          sftp.end && sftp.end()
-          resolve("exists")
-        } else {
-          resolve("does not exist")
-        }
-      })
-    })
-  })
-}
-
+/**
+ * @description This function uses SFTP to check if a file exists at the given remote path.
+ * @param {string} filePath - The remote path of the file to check
+ * @returns {string>} - Status of the file existence check: "exists", "does not exist", "sftp error", or "tunnel inactive"
+ */
 export async function checkRemoteFileExists(filePath) {
   // Ensure tunnel is active and SSH client is available
   console.log("checkRemoteFileExists", filePath)
@@ -353,6 +382,11 @@ export async function checkRemoteFileExists(filePath) {
   }
 }
 
+/**
+ * @description This function uses SFTP to call lstat on a remote file path.
+ * @param {string} filePath - The remote path of the file to check
+ * @returns {{ isDir: boolean, isFile: boolean, stats: Object } | string} - Returns an object with file stats or "sftp error" if an error occurs.
+ */
 export async function getRemoteLStat(filePath) {
   console.log("getRemoteLStat", filePath)
   // Ensure tunnel is active and SSH client is available
@@ -364,29 +398,33 @@ export async function getRemoteLStat(filePath) {
   }
     const getSftp = () => new Promise((resolve, reject) => {
     activeTunnel.sftp((err, sftp) => {
-      if (err) return reject(err);
-      resolve(sftp);
-    });
-  });
+      if (err) return reject(err)
+      resolve(sftp)
+    })
+  })
 
   const lstatFile = (sftp, filePath) => new Promise((resolve, reject) => {
     sftp.stat(filePath, (err, stats) => {
-      if (err) return reject(err); // File does not exist
-      resolve(stats);
-    });
-  });
+      if (err) return reject(err) // File does not exist
+      resolve(stats)
+    })
+  })
 
   try {
     const sftp = await getSftp()
     const fileStats = await lstatFile(sftp, filePath)
     sftp.end && sftp.end()
-    return fileStats
+    return { isDir: fileStats.isDirectory(), isFile: fileStats.isFile(), stats: fileStats }
   } catch (error) {
-    console.error("SFTP error:", error);
-    return "sftp error";
+    console.error("SFTP error:", error)
+    return "sftp error"
   }
 }
 
+/**
+ * @description This function uses a terminal command to detect the operating system of the remote server.
+ * @returns {Promise<string>}
+ */
 export async function detectRemoteOS() {
   return new Promise((resolve, reject) => {
     activeTunnel.exec("uname -s", (err, stream) => {
@@ -416,85 +454,422 @@ export async function detectRemoteOS() {
   })
 }
 
-// Unused
-export function getRemoteMongoDBPath() {
-  const remotePlatform = detectRemoteOS()
-
-  if (remotePlatform === "win32") {
-    // Check if mongod is in the process.env.PATH
-    const paths = process.env.PATH.split(path.delimiter)
-    for (let i = 0; i < paths.length; i++) {
-      const binPath = path.join(paths[i], "mongod.exe")
-      if (fs.existsSync(binPath)) {
-        console.log("mongod found in PATH")
-        return binPath
-      }
+/**
+ * @description This request handler creates a new remote folder in the specified parent path.
+ * @param {string} path - The parent path where the new folder will be created
+ * @param {string} folderName - The name of the new folder to be created
+ * @returns {Promise<{success: boolean, error?: string}>}
+ */
+ipcMain.handle('createRemoteFolder', async (_event, { path: parentPath, folderName }) => {
+  console.log('createRemoteFolder', parentPath, folderName)
+  const activeTunnel = getActiveTunnel()
+  // Helper to get SFTP client
+  function getSftp(cb) {
+    if (!activeTunnel) return cb(new Error('No active SSH tunnel'))
+    if (activeTunnel.sftp) {
+      return activeTunnel.sftp(cb)
+    } else if (activeTunnel.sshClient && activeTunnel.sshClient.sftp) {
+      return activeTunnel.sshClient.sftp(cb)
+    } else {
+      return cb(new Error('No SFTP available'))
     }
-    // Check if mongod is in the default installation path on Windows - C:\Program Files\MongoDB\Server\<version to establish>\bin\mongod.exe
-    const programFilesPath = process.env["ProgramFiles"]
-    if (programFilesPath) {
-      const mongoPath = path.join(programFilesPath, "MongoDB", "Server")
-      // Check if the MongoDB directory exists
-      if (!fs.existsSync(mongoPath)) {
-        console.error("MongoDB directory not found")
-        return null
-      }
-      const dirs = fs.readdirSync(mongoPath)
-      for (let i = 0; i < dirs.length; i++) {
-        const binPath = path.join(mongoPath, dirs[i], "bin", "mongod.exe")
-        if (fs.existsSync(binPath)) {
-          return binPath
-        }
-      }
-    }
-    console.error("mongod not found")
-    return null
-  } else if (process.platform === "darwin") {
-    // Check if it is installed in the .medomics directory
-    const binPath = path.join(process.env.HOME, ".medomics", "mongodb", "bin", "mongod")
-    if (fs.existsSync(binPath)) {
-      console.log("mongod found in .medomics directory")
-      return binPath
-    }
-    if (process.env.NODE_ENV !== "production") {
-      // Check if mongod is in the process.env.PATH
-      const paths = process.env.PATH.split(path.delimiter)
-      for (let i = 0; i < paths.length; i++) {
-        const binPath = path.join(paths[i], "mongod")
-        if (fs.existsSync(binPath)) {
-          console.log("mongod found in PATH")
-          return binPath
-        }
-      }
-      // Check if mongod is in the default installation path on macOS - /usr/local/bin/mongod
-      const binPath = "/usr/local/bin/mongod"
-      if (fs.existsSync(binPath)) {
-        return binPath
-      }
-    }
-    console.error("mongod not found")
-    return null
-  } else if (process.platform === "linux") {
-    // Check if mongod is in the process.env.PATH
-    const paths = process.env.PATH.split(path.delimiter)
-    for (let i = 0; i < paths.length; i++) {
-      const binPath = path.join(paths[i], "mongod")
-      if (fs.existsSync(binPath)) {
-        return binPath
-      }
-    }
-    console.error("mongod not found in PATH" + paths)
-    // Check if mongod is in the default installation path on Linux - /usr/bin/mongod
-    if (fs.existsSync("/usr/bin/mongod")) {
-      return "/usr/bin/mongod"
-    }
-    console.error("mongod not found in /usr/bin/mongod")
-
-    if (fs.existsSync("/home/" + process.env.USER + "/.medomics/mongodb/bin/mongod")) {
-      return "/home/" + process.env.USER + "/.medomics/mongodb/bin/mongod"
-    }
-    return null
-  } else {
-    return "mongod"
   }
-}
+  // Normalize path for SFTP: always use absolute, default to home dir as '.'
+  function normalizePath(p) {
+    if (!p || p === '') return '.'
+    if (p === '~') return '.'
+    if (p.startsWith('~/')) return p.replace(/^~\//, '')
+    return p
+  }
+  return new Promise((resolve) => {
+    getSftp(async (err, sftp) => {
+      if (err) return resolve({ success: false, error: err.message })
+      let sftpClosed = false
+      function closeSftp() {
+        if (sftp && !sftpClosed) {
+          if (typeof sftp.end === 'function') {
+            try { sftp.end() } catch (e) {}
+          } else if (typeof sftp.close === 'function') {
+            try { sftp.close() } catch (e) {}
+          }
+          sftpClosed = true
+        }
+      }
+      try {
+        console.log('Creating folder', folderName, 'in', parentPath)
+        const parent = normalizePath(parentPath)
+        // Step 1: resolve canonical parent path
+        let canonicalParent = await new Promise((res, rej) => {
+          sftp.realpath(parent, (e, abs) => e ? res(parent) : res(abs))
+        })
+        // Step 2: build new folder path
+        let newFolderPath = folderName ? canonicalParent.replace(/\/$/, '') + '/' + folderName : canonicalParent
+        // Step 3: create directory
+        console.log('Resolved new folder path:', newFolderPath)
+        await new Promise((res, rej) => {
+          sftp.mkdir(newFolderPath, (e) => e ? rej(e) : res())
+        })
+        closeSftp()
+        console.log('Folder created successfully')
+        resolve({ success: true })
+      } catch (e) {
+        closeSftp()
+        console.error('Error creating remote folder:', e)
+        resolve({ success: false, error: e.message })
+      }
+    })
+  })
+})
+
+
+/**
+ * @description This request handler manages the remote navigation of folders on the server.
+ * @param {string} action - 'list' to display files and folders, 'up' to go back a directory or 'into' to enter it
+ * @param {string} path - The remote path to navigate
+ * @param {string} dirName - The name of the directory to enter (only used for 'into' action)
+ * @returns {Promise<{success: boolean, error?: string}>}
+ */
+ipcMain.handle('navigateRemoteDirectory', async (_event, { action, path: currentPath, dirName }) => {
+  const activeTunnel = getActiveTunnel()
+  // Helper to get SFTP client
+  function getSftp(cb) {
+    if (!activeTunnel) return cb(new Error('No active SSH tunnel'))
+    if (activeTunnel.sftp) {
+      // ssh2 v1.15+ attaches sftp method directly
+      return activeTunnel.sftp(cb)
+    } else if (activeTunnel.sshClient && activeTunnel.sshClient.sftp) {
+      return activeTunnel.sshClient.sftp(cb)
+    } else {
+      return cb(new Error('No SFTP available'))
+    }
+  }
+
+  // Promisified SFTP realpath
+  function sftpRealpath(sftp, p) {
+    return new Promise((resolve, reject) => {
+      sftp.realpath(p, (err, absPath) => {
+        if (err) return reject(err)
+        resolve(absPath)
+      })
+    })
+  }
+
+  // Promisified SFTP readdir
+  function sftpReaddir(sftp, p) {
+    return new Promise((resolve, reject) => {
+      sftp.readdir(p, (err, list) => {
+        if (err) return reject(err)
+        resolve(list)
+      })
+    })
+  }
+
+  // Normalize path for SFTP: always use absolute, default to home dir as '.'
+  function normalizePath(p) {
+    if (!p || p === '') return '.' // SFTP: '.' means home dir
+    if (p === '~') return '.'
+    if (p.startsWith('~/')) return p.replace(/^~\//, '')
+    return p
+  }
+
+  return new Promise((resolve) => {
+    getSftp(async (err, sftp) => {
+      if (err) return resolve({ path: currentPath, contents: [], error: err.message })
+      let targetPath = normalizePath(currentPath)
+      let sftpClosed = false
+      // Helper to close SFTP session safely
+      function closeSftp() {
+        if (sftp && !sftpClosed) {
+          if (typeof sftp.end === 'function') {
+            try { sftp.end() } catch (e) {}
+          } else if (typeof sftp.close === 'function') {
+            try { sftp.close() } catch (e) {}
+          }
+          sftpClosed = true
+        }
+      }
+      try {
+        // Step 1: resolve canonical path (absolute)
+        let canonicalPath = await sftpRealpath(sftp, targetPath).catch(() => targetPath)
+        // Step 2: handle navigation action
+        if (action === 'up') {
+          // Go up one directory
+          if (canonicalPath === '/' || canonicalPath === '' || canonicalPath === '.') {
+            // Already at root/home
+            // List current
+          } else {
+            let parts = canonicalPath.split('/').filter(Boolean)
+            if (parts.length > 1) {
+              parts.pop()
+              canonicalPath = '/' + parts.join('/')
+            } else {
+              canonicalPath = '/'
+            }
+          }
+        } else if (action === 'into' && dirName) {
+          // Always join using absolute path
+          if (canonicalPath === '/' || canonicalPath === '') {
+            canonicalPath = '/' + dirName
+          } else if (canonicalPath === '.') {
+            // Home dir: get its absolute path
+            canonicalPath = await sftpRealpath(sftp, '.').catch(() => '/')
+            canonicalPath = canonicalPath.replace(/\/$/, '') + '/' + dirName
+          } else {
+            canonicalPath = canonicalPath.replace(/\/$/, '') + '/' + dirName
+          }
+          // Re-resolve in case of symlinks
+          canonicalPath = await sftpRealpath(sftp, canonicalPath).catch(() => canonicalPath)
+        } else if (action === 'list') {
+          // Just list current
+        }
+        // Step 3: list directory
+        let entries = await sftpReaddir(sftp, canonicalPath).catch(() => [])
+        let contents = Array.isArray(entries)
+          ? entries.filter(e => e.filename !== '.' && e.filename !== '..').map(e => ({
+              name: e.filename,
+              type: e.attrs.isDirectory() ? 'dir' : 'file'
+            }))
+          : []
+        closeSftp()
+        resolve({ path: canonicalPath, contents })
+      } catch (e) {
+        closeSftp()
+        resolve({ path: currentPath, contents: [], error: e.message })
+      }
+    })
+  })
+})
+
+ipcMain.handle('startSSHTunnel', async (_event, params) => {
+  return startSSHTunnel(params)
+});
+
+ipcMain.handle('startMongoTunnel', async () => {
+  return startMongoTunnel()
+});
+
+ipcMain.handle('confirmMongoTunnel', async () => {
+  return confirmMongoTunnel()
+});
+
+ipcMain.handle('stopSSHTunnel', async () => {
+  return stopSSHTunnel()
+});
+
+ipcMain.handle('getRemoteLStat', async (_event, path) => {
+  return getRemoteLStat(path)
+})
+
+ipcMain.handle('checkRemoteFileExists', async (_event, path) => {
+  return checkRemoteFileExists(path)
+})
+
+ipcMain.handle('setRemoteWorkspacePath', async (_event, path) => {
+  return setRemoteWorkspacePath(path)
+})
+
+/**
+ * @description This request handler lists the contents of a remote directory on the server.
+ * @param {string} path - The remote path of the folder to list
+ * @returns {Promise<{success: boolean, error?: string}>}
+ */
+ipcMain.handle('listRemoteDirectory', async (_event, { path: remotePath }) => {
+  return new Promise((resolve, reject) => {
+    const activeTunnel = getActiveTunnel()
+    if (!activeTunnel) {
+      return resolve({ path: remotePath, contents: [], error: 'No active SSH tunnel' })
+    }
+    try {
+      activeTunnel.sftp((err, sftp) => {
+        if (err || !sftp) return resolve({ path: remotePath, contents: [], error: err ? err.message : 'No SFTP' })
+        // Normalize path for SFTP: always use absolute, default to home dir as '.'
+        function normalizePath(p) {
+          if (!p || p === '') return '.' // SFTP: '.' means home dir
+          if (p === '~') return '.'
+          if (p.startsWith('~/')) return p.replace(/^~\//, '')
+          return p
+        }
+        const targetPath = normalizePath(remotePath)
+        // First, resolve canonical/absolute path
+        sftp.realpath(targetPath, (err2, absPath) => {
+          const canonicalPath = (!err2 && absPath) ? absPath : targetPath
+          sftp.readdir(canonicalPath, (err3, list) => {
+            // Always close SFTP session after use
+            if (sftp && typeof sftp.end === 'function') {
+              try { sftp.end() } catch (e) {}
+            } else if (sftp && typeof sftp.close === 'function') {
+              try { sftp.close() } catch (e) {}
+            }
+            if (err3) return resolve({ path: canonicalPath, contents: [], error: err3.message })
+            const contents = Array.isArray(list)
+              ? list.filter(e => e.filename !== '.' && e.filename !== '..').map(e => ({
+                  name: e.filename,
+                  type: e.attrs.isDirectory() ? 'dir' : 'file'
+                }))
+              : []
+            resolve({ path: canonicalPath, contents })
+          })
+        })
+      })
+    } catch (e) {
+      resolve({ path: remotePath, contents: [], error: e.message })
+    }
+  })
+})
+
+// SSH key management
+ipcMain.handle('generateSSHKey', async (_event, { comment, username }) => {
+  try {
+    const userDataPath = app.getPath('userData')
+    const privKeyPath = path.join(userDataPath, `${username || 'user'}_id_rsa`)
+    const pubKeyPath = path.join(userDataPath, `${username || 'user'}_id_rsa.pub`)
+    let privateKey, publicKey
+    if (fs.existsSync(privKeyPath) && fs.existsSync(pubKeyPath)) {
+      privateKey = fs.readFileSync(privKeyPath, 'utf8')
+      publicKey = fs.readFileSync(pubKeyPath, 'utf8')
+    } else {
+      const result = await generateSSHKeyPair(comment, username)
+      privateKey = result.privateKey
+      publicKey = result.publicKey
+      fs.writeFileSync(privKeyPath, privateKey, { mode: 0o600 })
+      fs.writeFileSync(pubKeyPath, publicKey, { mode: 0o644 })
+    }
+    return { privateKey, publicKey }
+  } catch (err) {
+    return { error: err.message }
+  }
+})
+
+ipcMain.handle('getSSHKey', async (_event, { username }) => {
+  try {
+    const userDataPath = app.getPath('userData')
+    const privKeyPath = path.join(userDataPath, `${username || 'user'}_id_rsa`)
+    const pubKeyPath = path.join(userDataPath, `${username || 'user'}_id_rsa.pub`)
+    let privateKey, publicKey
+    if (fs.existsSync(privKeyPath) && fs.existsSync(pubKeyPath)) {
+      privateKey = fs.readFileSync(privKeyPath, 'utf8')
+      publicKey = fs.readFileSync(pubKeyPath, 'utf8')
+      return { privateKey, publicKey }
+    } else {
+      return { privateKey: '', publicKey: '' }
+    }
+  } catch (err) {
+    return { error: err.message }
+  }
+})
+
+
+
+//  ----- Unused -----
+// export function getRemoteMongoDBPath() {
+//   const remotePlatform = detectRemoteOS()
+
+//   if (remotePlatform === "win32") {
+//     // Check if mongod is in the process.env.PATH
+//     const paths = process.env.PATH.split(path.delimiter)
+//     for (let i = 0; i < paths.length; i++) {
+//       const binPath = path.join(paths[i], "mongod.exe")
+//       if (fs.existsSync(binPath)) {
+//         console.log("mongod found in PATH")
+//         return binPath
+//       }
+//     }
+//     // Check if mongod is in the default installation path on Windows - C:\Program Files\MongoDB\Server\<version to establish>\bin\mongod.exe
+//     const programFilesPath = process.env["ProgramFiles"]
+//     if (programFilesPath) {
+//       const mongoPath = path.join(programFilesPath, "MongoDB", "Server")
+//       // Check if the MongoDB directory exists
+//       if (!fs.existsSync(mongoPath)) {
+//         console.error("MongoDB directory not found")
+//         return null
+//       }
+//       const dirs = fs.readdirSync(mongoPath)
+//       for (let i = 0; i < dirs.length; i++) {
+//         const binPath = path.join(mongoPath, dirs[i], "bin", "mongod.exe")
+//         if (fs.existsSync(binPath)) {
+//           return binPath
+//         }
+//       }
+//     }
+//     console.error("mongod not found")
+//     return null
+//   } else if (process.platform === "darwin") {
+//     // Check if it is installed in the .medomics directory
+//     const binPath = path.join(process.env.HOME, ".medomics", "mongodb", "bin", "mongod")
+//     if (fs.existsSync(binPath)) {
+//       console.log("mongod found in .medomics directory")
+//       return binPath
+//     }
+//     if (process.env.NODE_ENV !== "production") {
+//       // Check if mongod is in the process.env.PATH
+//       const paths = process.env.PATH.split(path.delimiter)
+//       for (let i = 0; i < paths.length; i++) {
+//         const binPath = path.join(paths[i], "mongod")
+//         if (fs.existsSync(binPath)) {
+//           console.log("mongod found in PATH")
+//           return binPath
+//         }
+//       }
+//       // Check if mongod is in the default installation path on macOS - /usr/local/bin/mongod
+//       const binPath = "/usr/local/bin/mongod"
+//       if (fs.existsSync(binPath)) {
+//         return binPath
+//       }
+//     }
+//     console.error("mongod not found")
+//     return null
+//   } else if (process.platform === "linux") {
+//     // Check if mongod is in the process.env.PATH
+//     const paths = process.env.PATH.split(path.delimiter)
+//     for (let i = 0; i < paths.length; i++) {
+//       const binPath = path.join(paths[i], "mongod")
+//       if (fs.existsSync(binPath)) {
+//         return binPath
+//       }
+//     }
+//     console.error("mongod not found in PATH" + paths)
+//     // Check if mongod is in the default installation path on Linux - /usr/bin/mongod
+//     if (fs.existsSync("/usr/bin/mongod")) {
+//       return "/usr/bin/mongod"
+//     }
+//     console.error("mongod not found in /usr/bin/mongod")
+
+//     if (fs.existsSync("/home/" + process.env.USER + "/.medomics/mongodb/bin/mongod")) {
+//       return "/home/" + process.env.USER + "/.medomics/mongodb/bin/mongod"
+//     }
+//     return null
+//   } else {
+//     return "mongod"
+//   }
+// }
+
+// export function checkRemoteFolderExists(folderPath) {
+//   // Ensure tunnel is active and SSH client is available
+//   const tunnelObject = getActiveTunnel()
+//   if (!tunnelObject) {
+//     const errMsg = "No active SSH tunnel for remote folder creation."
+//     console.error(errMsg)
+//     return Promise.resolve("tunnel inactive")
+//   }
+
+//   return new Promise((resolve, reject) => {
+//     tunnelObject.sftp((err, sftp) => {
+//       if (err) {
+//         console.error("SFTP error:", err)
+//         resolve("sftp error")
+//         return
+//       }
+
+//       // Check if folder exists
+//       sftp.stat(folderPath, (statErr, stats) => {
+//         if (!statErr && stats && stats.isDirectory()) {
+//           // Folder exists
+//           sftp.end && sftp.end()
+//           resolve("exists")
+//         } else {
+//           resolve("does not exist")
+//         }
+//       })
+//     })
+//   })
+// }
